Add leaderboard both mode (masih ada error)

+++ Assets/Scripts/UI/Score.cs
@@ -7,9 +7,9 @@ using System;
 public class Score 
 {
     public string name;
-    public int score;
+    public float score;
 
-    public Score(string name, int score)
+    public Score(string name, float score)
     {
         this.name = name;
         this.score = score;
